fix: patch Requests page crash, scope bypass, and notification isolation
  - Fix Pydantic v2 Decimal serialization: quantity_requested typed as float
    in InventoryRequestOut so the API returns a JSON number instead of a string,
    preventing the "toFixed is not a function" crash that blanked the page
  - Add ErrorBoundary around the app shell so future render errors show a message
    instead of a blank screen
  - Enforce group-lead scoping on web approve/deny endpoints: scope check via
    check_request_scope() runs before any mutation; 403 on cross-group access
  - Enforce same group-lead scoping in Telegram /approve, /deny, and /requests
  - Isolate post-commit Telegram notifications in try/except so a notification
    failure never masks a successful approval or denial already committed to DB
  - Move all request mutation query invalidations to onSettled so the list
    always refetches regardless of whether the mutation succeeded or errored
  - Add 23 backend tests covering approval lifecycle, group-lead scope enforcement,
    cross-group rejection, admin global access, and notification failure isolation

+++ backend/app/schemas/inventory_request.py
@@ -37,7 +37,7 @@ class InventoryRequestOut(BaseModel):
     approver_id: Optional[int] = None
     item_id: Optional[int] = None
     bin_id: Optional[int] = None
-    quantity_requested: Decimal
+    quantity_requested: float
     status: str
     reason: Optional[str] = None
     denial_reason: Optional[str] = None
